server/requirements.txt
Flask
gunicorn
flask-cors
google-cloud-firestore